GSL: check the RSP and unmap the memory if needed
Check the RSP token from spf, since the timeout issue may
happened previously which will result in token not matched
with the sent packet.

If the delayed packet or unconsumed packet is a memory map
command, cache the signal then unmap it in proper time.
If the delayed packet is a common command, just skip it and
continue to wait for the right rsp.

Signed-off-by: Pratyush Meduri <mpratyus@qti.qualcomm.com>