gsl: skip unsuccessful PD service registration for a subsystem
PD notifier registration for CDSP service is failing
because CDSP is not present on some variant.This, in
turn, is causing GSL initialization failure.

Ideally, the servreg domain list could have indicated
this, but it may not be possible due to cross-platform
build dependencies.Therefore, ignore the error to
prevent GSL initialization failure.

Signed-off-by: Pratyush Meduri <mpratyus@qti.qualcomm.com>